add bokeh.testing.selenium
