+++ tests/integration/interaction/test_datatable.py
@@ -11,7 +11,7 @@ from bokeh.models import (
 )
 from bokeh.layouts import column
 from selenium.webdriver.common.action_chains import ActionChains
-from tests.integration.utils import has_no_console_errors
+from bokeh.testing.selenium import has_no_console_errors
 
 import pytest
 pytestmark = pytest.mark.integration
