+++ tests/integration/interaction/test_patches_preselection.py
@@ -8,7 +8,7 @@ from bokeh.models import (
     Range1d,
 )
 from selenium.webdriver.common.action_chains import ActionChains
-from tests.integration.utils import has_no_console_errors
+from bokeh.testing.selenium import has_no_console_errors
 
 import pytest
 pytestmark = pytest.mark.integration
